migrate BulkEditView component to TypeScript
